feat: Vectorised indexing for diagnostic values
The previous implementation would retrieve the diagnostic value based on the index value stored within the row of the dataframe. When watching the dask distributed dashbaord, I noticed the computation would get stuck on a couple of task never progressing until it magically figures it out. By using the dask.distributed.progress and dask.distributed.print, I managed to narrow down where this bottleneck was occuring.

The bottleneck was in the retrieval of the computed diagnostic value. This location corresponds to the mysterious failure of processing the NOAA data on the HPC. The new implementation uses the identified lat, lon, level and time index to index directly into the computed diagnostic array instead of going by row. As dask does not support the multi-dimensional advanced indexing that's in numpy, the equivalent flattened index is found. da.ravel() is used to create a view (not a copy that flattened would give) such that the flattened index can be used to access the values accordingly