fix: Remove calls to blocking_wait_future
In #32, I had mistaken thought that the calls to this function were what fixed it, I was wrong. It had actually been the calls to persist() that were fixing it. When persist() is invoked, it creates a future pointing to that collection. This indicates to the client to hold it in memory (see https://distributed.dask.org/en/stable/memory.html?highlight=futures). Thus, waiting on the futures to complete serves no purpose