fix: Incorrectly implemented BrownIndex1 equation
The bug was highlighted due to the RuntimeWarning: invalid value encountered in sqrt. As this is caused by taking the sqrt of a negative number and most of the diagnostics are always positive, it hinted at an implementation error. Upon inspection of the equationa nd what was implemented, it became clear that the terms were all missing a square