feat: Join on index to get a GeoPandas Dataframe without needing to convert from a dataframe
See the comments in _retrieve_grouped_columns. This updated implementation means that join is call on the grid GeoSeries to make the product a Geopandas GeoDataFrame instead of doing the join on the data frame then converting into a GeoDataFrame. The latter would need to manually set the crs and do the conversion. However, this updated implementaiton has the cost of a set_index