fix: Add fallback version to hatch version
When the .git/ folder is not present, hatch is unable to figure out what the version is as no VCS is present. This causes uv sync --all-groups to fail with the LookupError: Error getting the version from source `vcs`: setuptools-scm was unable to detect version. By providing a fall back version, hatch can use that when the vcs is not detected