feat: Improve representation of grid boxes with an insufficient amount of observations
In the existing plots, there are two ways in an AUC value of 0 occurs. The first is where there are no positive observations of turbulence. The second is where there is only a single observation at that point. This makes it difficult to interpret the information presented in the plot as a single value maps to multiple meanings. I had been slowly arriving at this conclusion but showing the plots to Ed and Rafa confirmed it.

How it works? The AUC calculation in the group by operation will return np.nan when the minimum group size is not met. When plotting the gv.Polygons, the colour bar is configured such that values are limited to the range of (0, 1), i.e. valid values of AUC, and any value that is NaN will have a different grey colour to indicate that it is inactive/irrelevant