refactor: Move computing of binary classification rates into a function within metrics
Part of a larger refactor. This is to reduce the code duplication that would be present for computing AUC and TSS.

This refacotring also allows testing. One test is to check that the two binary classification methods for ROC variables are equivalent. The second is to ensure the edge cases are hit and the value errors are raised