feat: Make aggregation more efficient
I believe that by applying the condition at the series level on the dataframe, it would be more efficient than during the groupby aggregation. This also means that the builtin pd.SeriesGroupBy.any() method can be used. As the pandas docs state that their built in functions are more efficient than user defined ones, I'm left to believe this should be more efficient. Plus, it reads a bit better