fix: Ensure dask arrays of dd.Series have the same number of chunks to prevent re-chunking
When running on a larger dataset, e.g. on the HPC, there is a warning in the logs - `PerformanceWarning: Increasing number of chunks by factor of`. This is due to the rechunking of a dask array.

From reading various dask discussions, it appears that this rechunking occurs during invocations of blockwise or map_overlap or map_blocks method. This rechunking is to ensure that the input arrays have the same chunks such that they can be successfully iterated over.

As I know that my arrays will be the same length, it is desireable to control this at the point they are being created. Thus, the to_dask_array() function is used instead of the .values.compute_chunk_sizes() as the former allows one to pass the length of each chunk as inputs.