fix: Pressure level coordinate was being converted to altitudes due to in place manipulation of arrays
Due to the implementation of the pressure_to_altitude_icao method and the shallow copies of the arrays that were being passed to the function, the pressure levels were being modified in place instead of on a copy leading to the pressure level coordinate having the same values as the altitudes. This caused issues with the calculation of the turbulence diagnostics and when the observational data was being compared against them.