improvement: Convert row-wise apply to be on sections blocks of dataframe when converting altitude to closest pressure level
Using dd.apply requires iterating over each row of the dataframe, this makes it extremely slow. This converts the row-wise operation to being on each parition of the dataframe.

This was introduced as running a simple test to identify the turbulence frequency was taking painfully long and the process would randomly die around this part of the code. From the dask dashboard, it was also clear that the operaiton that was being performed was not being parallelised and required the usage of a large amount of memory.

With the introduction of this implemntation, the code went from failing after > 1.5hrs before the first iteration to finishing the first iteration in < 4mins