improvement: Removed harmonisation code path
This code path doesn't really work :( It is very slow and the invocation to expand the items in the geometry column into min_lon, min_lat, etc, was slowing down the code path that does work. Thus, I decided this was the right juncture to remove this essentially dead code path.

This only removes the invocations within the launcher and beyond. It does not touch the outward facing configuraiton. That will be done in the next commit.