deps: Added sparse as a dependency
From the geniuses in scikit-learn, a very simple way to calculate the binary confusion matrix is through the usage of sparse arrays. Their implementation uses the scipy sparse. However, due to the size of my true and pred arrays, I will need them to be dask arrays. Hence, sparse has been added as dependency as it is dask compatible and works in the same way from the small experiment I ran