feat: Implemented method to compute confusion matrix
This uses the genius method implemented in scikit-learn. It leverages the fact that sparse arrays can be defined Coordinate Format (COO). As the definition of it would include the shape ((2,2) for binary classification), the pairs of coordinates of (ground_truth, prediction) would specify which quadrant a given data point falls into!

When the array is converted to dense, the sparse array simply sums up the number that fall into each category to give you the number of true positives, false positives, etc! It is super smart!!