feat: Implemented a vectorised method to compute matthews correlation coefficient using bitwise operations
The original method stolen from sciki-learn only works on a 1D array. Attempting to get it to work on multiple dimensions using xr.apply_ufunc or numba vectorisation didn't pan out. When looking at the theory for conditional probability to check that my implementation for another function was correct, I realised that I could use bitwise operations to compute the terms of the matthews correlation coefficient as it all boils down to the conditional probability table.