feat: Implemented method to find parent region of points which intersect
This method uses the boolean mask of the points which intersect to find the corresponding value in the labeled array. As all regions will have the same number, the unique values found from the mask can be used to find all values which match the identified regions.

Without the labeling, it will not be possible to find the parent region as the intersection would cause the information to be lost and there are an infinite number of permutations that could've resulted in that intersection pattern. Thus, by first adding the information about which region a point belongs to, it's possible to recover the parent regions