feat: Make e_sat_ice a numba ufunc
The implementation of the method uses numpy ufuncs log and exp. Since these are then multiplied by constants, I figured this could easily be converted into a ufunc on it's own to make the code slightly more efficient.

Also, turned the **2 to x * x to make it more efficient in python. I don't know if this would make a difference now that it's a numba function