fix: Condition spuriously fails on HPC. Data already in decreasing so this should never fail (#72)
On the HPC, this condition will fail on one specific configuration. As I'm not able to replicate the HPC set up locally, I've not been able to test it. However, I chanced upon a situaiton where this was replicated and found that for that case, flipping the condition fixed the issue.