feat: Finds coordinate on great circle between two points
NOT TESTED!!! This finds coordinates on the great circle between two points. The number of points between the two is estimated from the grid size. It has been set so that it is approximately similar to the original grid. However, this can be tweaked through the n_points_safety_factor kwarg.