feat: Added another standard cmap and allow all to be resampled
This reworks the way get_a_default_cmap works to make it more flexible. firstly, it allows kwargs to be passed to the load_cmap function. secondly, it allows all smaps to be resample to the specified value. lastly, it adds a default cmap for the distances between plots feature