feat: Composite an array about a unique jet core
The goal is to perform a composite about the jet core. To simplify things, this implementation only considers the case where there is a single jet core for a given longitude. It also assume that the data is for a single hemisphere.

I attempted to do this operation using xarray operations only. However, I got stumpted at the indexing to map the values and using np.take_along_axis to work as intended. As such, I surmised that the next best thing is to do as much as possible at the xarray level, then for the finicky bit of copying data over appropriately, use guvectorise