feat: NOT TESTED Used sparse.COO to regrid AMADAR observations
OMG THIS RUNS!!!! I did not expect it to work so easily. I have not checked the correctness and it still need a lot of cleaning up on the code especially with the repeated code.

This uses map_blocks to map the various blocks from the dataframe into a 4D array. COO is used as I can define the coordinate for each data point. If the data point is repeated, i.e. true twice, then I get the frequency of the turbulence observed at that data point for that point in space and time!