fix: Bug in original implementation meant that each chunk was a unique COO array
When testing with data that has been chunked, I found a bug that meant: 1) calling compute would break due to the computed result having n_chunks*chunk_size_of_time_axis in for the size of the time axis, and 2) invoking sum(dim='time') would have the correct indices modified but the value would be n_chunks*count.

Through the debugging process, I eventually figured out that I'd have no choice but to use map_blocks to instantiate the sparse.COO instance then use da.from_array to convert it back to daks. This might negate the benefits of using dask as the data has been converted out of the dask system and is loaded as raw values into dask again. So, my plan is to now whack it with a large data to see if a large graph is sent when operations are performed on it