MAINT: Update all dependencies (#126)
* maint: Updated all dependencies

* maint: Add basedpyright as dep for IDE

* fix: Fix type issue after upgrading packages (xarray)

* test: Attempt to fix failing test in CI

For some reason, the tests fail in the CI with a h5 related error, so I'm going to try fixing it by changing the default engine. I can't replicate this failure mode locally.

* test: Relax test to ensure it passes consistently

When I run this test on my mac, there are no decmial issues. When run on hurry-up.vm, it fails due to a difference in the 6th decimal place